--- pkg/coded_err/codes.go
@@ -1,8 +1,0 @@
-package coded_err
-
-const (
-	CodeBatchSize        = "batch_size"
-	CodeConcurrencyLimit = "concurrency_limit"
-	CodeConfigInvalid    = "config_invalid"
-	CodeUnknown          = "unknown"
-)
