+++ pkg/inngest/batch.go
@@ -5,8 +5,8 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/inngest/inngest/pkg/coded_err"
 	"github.com/inngest/inngest/pkg/consts"
+	"github.com/inngest/inngest/pkg/syscode"
 )
 
 func NewEventBatchConfig(conf map[string]any) (*EventBatchConfig, error) {
@@ -63,8 +63,8 @@ func (c EventBatchConfig) IsEnabled() bool {
 
 func (c EventBatchConfig) IsValid() error {
 	if c.MaxSize < 2 {
-		return coded_err.Error{
-			Code:    coded_err.CodeBatchSize,
+		return syscode.Error{
+			Code:    syscode.CodeBatchSizeInvalid,
 			Message: fmt.Sprintf("batch size cannot be smaller than 2: %d", c.MaxSize),
 		}
 	}
