+++ pkg/inngest/concurrency.go
@@ -9,10 +9,10 @@ import (
 
 	"github.com/cespare/xxhash/v2"
 	"github.com/google/uuid"
-	"github.com/inngest/inngest/pkg/coded_err"
 	"github.com/inngest/inngest/pkg/consts"
 	"github.com/inngest/inngest/pkg/enums"
 	"github.com/inngest/inngest/pkg/expressions"
+	"github.com/inngest/inngest/pkg/syscode"
 )
 
 // ConcurrencyLimits represents concurrency limits specified for a function.
@@ -37,8 +37,8 @@ func (c ConcurrencyLimits) PartitionConcurrency() int {
 
 func (c ConcurrencyLimits) Validate(ctx context.Context) error {
 	if len(c.Limits) > consts.MaxConcurrencyLimits {
-		return coded_err.Error{
-			Code:    coded_err.CodeConcurrencyLimit,
+		return syscode.Error{
+			Code:    syscode.CodeConcurrencyLimitInvalid,
 			Message: fmt.Sprintf("There are more concurrency limits specified than the allowed max of: %d", consts.MaxConcurrencyLimits),
 		}
 	}
