+++ pkg/syscode/codes.go
@@ -1,0 +1,8 @@
+package syscode
+
+const (
+	CodeBatchSizeInvalid        = "batch_size_invalid"
+	CodeConcurrencyLimitInvalid = "concurrency_limit_invalid"
+	CodeConfigInvalid           = "config_invalid"
+	CodeUnknown                 = "unknown"
+)
